requirements.txt
Flask
Flask-Cors
Flask-RESTful
Flask-JWT-Extended
PyMongo
google-generativeai